Moved the tests in the correct fdlder
